% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfoCookie.R
\name{bioinfoCookie}
\alias{bioinfoCookie}
\title{Get bioinfo cookie}
\usage{
bioinfoCookie(force.refresh = FALSE)
}
\arguments{
\item{force.refresh}{Logical. If set to \code{TRUE}, cached cookie is not used and a new cookie is fetched.}
}
\value{
A character string representing the cookie
}
\description{
Get bioinfo cookie
}
\details{
User name is obtained by querying the operating system.

The function first checks whether a cached cookie exists: if so, it
simply returns it without querying the webserver; otherwise, it
fetches a newly backed cookie from the webserver.
}
\examples{
\dontrun{bioinfoCookie()}
}
\author{
Jitao David Zhang, Detlef Wolf, and Clemens Brogers
}

